Rename state_snapshot to util module.
